+++ grid2op/Chronics/changeNothing.py
@@ -57,14 +57,14 @@ class ChangeNothing(GridValue):
                            action_class=TopologyAndDispatchAction)
         
     """
-
+    MULTI_CHRONICS = False
     def __init__(
         self,
         time_interval=timedelta(minutes=5),
         max_iter=-1,
         start_datetime=datetime(year=2019, month=1, day=1),
         chunk_size=None,
-        **kargs
+        **kwargs
     ):
         GridValue.__init__(
             self,